refactor(core): move EnvironmentActorPool to core/distributed.py (from utils/ray.py)
Distributed rollout is domain infra (it knows Environment/Action/StepResult), not
a generic util, so it belongs in core/ — purpose-named, with Ray as an internal
detail. Ray stays optional/runtime-provided: it's imported lazily (CLI distributed
path) and under TYPE_CHECKING (evaluator), and core/__init__ never imports it, so
`import strands_env.core` does not pull Ray.
